# api/openapi/tars-mvp.yaml
openapi: 3.1.0
info:
  title: TARS MVP API
  version: 0.1.0
  description: |
    Draft API definition aligned with tars_technical_design.md v1.4.
    Tag names describe API contract groupings, not Web Console page ownership.
    In particular, the `ops` tag means authenticated control-plane APIs and is
    not equivalent to the `/ops` UI page; some `ops`-tagged APIs are primarily
    surfaced from other daily object pages such as `/identity/*`, `/providers`,
    and `/connectors`.
servers:
  - url: https://tars.example.com
...
paths:
...
  /api/v1/automations:
    get:
      tags: [ops]
      summary: List governed automation jobs
      operationId: listAutomations
      security:
        - OpsBearer: []
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [skill, connector_capability]
        - name: status
          in: query
          schema:
            type: string
        - name: enabled
          in: query
          schema:
            type: boolean
        - name: q
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
        - name: sort_by
          in: query
          schema:
            type: string
            enum: [display_name, next_run_at, last_run_at, status, type]
        - name: sort_order
          in: query
          schema:
            type: string
            enum: [asc, desc]
      responses:
        "200":
          description: Automation job list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AutomationListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/OpsAPIDisabled"
        "409":
          description: Automation manager not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    post:
      tags: [ops]
      summary: Create or upsert an automation job
      operationId: createAutomation
      security:
        - OpsBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AutomationUpsertRequest"
      responses:
        "201":
          description: Automation job created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AutomationJob"
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/OpsAPIDisabled"
        "409":
          description: Automation manager not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /api/v1/automations/{job_id}:
    get:
      tags: [ops]
      summary: Get automation job detail
      operationId: getAutomation
      security:
        - OpsBearer: []
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Automation job detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AutomationJob"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Automation manager not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    put:
      tags: [ops]
      summary: Update an automation job
      operationId: updateAutomation
      security:
        - OpsBearer: []
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AutomationUpsertRequest"
      responses:
        "200":
          description: Updated automation job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AutomationJob"
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Automation manager not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /api/v1/automations/{job_id}/enable:
    post:
      tags: [ops]
      summary: Enable an automation job
      operationId: enableAutomation
      security:
        - OpsBearer: []
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Enabled automation job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AutomationJob"
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/automations/{job_id}/disable:
    post:
      tags: [ops]
      summary: Disable an automation job
      operationId: disableAutomation
      security:
        - OpsBearer: []
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Disabled automation job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AutomationJob"
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/automations/{job_id}/run:
    post:
      tags: [ops]
      summary: Trigger an automation job immediately
      operationId: runAutomationNow
      security:
        - OpsBearer: []
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Automation job with latest run result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AutomationJob"
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    OpsBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque
  responses:
    Unauthorized:
      description: Authentication required or token invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Forbidden:
      description: Authenticated but not permitted
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    ValidationFailed:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    OpsAPIDisabled:
      description: Ops API disabled
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
...
    ListPage:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        has_next:
          type: boolean
        q:
          type: string
        sort_by:
          type: string
        sort_order:
          type: string
...
    AutomationSkillTarget:
      type: object
      required: [skill_id]
      properties:
        skill_id:
          type: string
        context:
          type: object
          additionalProperties: true
    AutomationConnectorCapability:
      type: object
      required: [connector_id, capability_id]
      properties:
        connector_id:
          type: string
        capability_id:
          type: string
        params:
          type: object
          additionalProperties: true
    AutomationRun:
      type: object
      required: [run_id, job_id, trigger, status]
      properties:
        run_id:
          type: string
        job_id:
          type: string
        trigger:
          type: string
        status:
          type: string
          enum: [running, completed, failed, blocked]
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
        attempt_count:
          type: integer
        summary:
          type: string
        error:
          type: string
        metadata:
          type: object
          additionalProperties: true
    AutomationJobState:
      type: object
      properties:
        status:
          type: string
          enum: [enabled, disabled, running, completed, failed, blocked]
        last_run_at:
          type: string
          format: date-time
        next_run_at:
          type: string
          format: date-time
        last_outcome:
          type: string
        last_error:
          type: string
        consecutive_failures:
          type: integer
        runs:
          type: array
          items:
            $ref: "#/components/schemas/AutomationRun"
        updated_at:
          type: string
          format: date-time
    AutomationJob:
      type: object
      required: [id, display_name, type, target_ref, schedule, enabled]
      properties:
        id:
          type: string
        display_name:
          type: string
        description:
          type: string
        agent_role_id:
          type: string
        type:
          type: string
          enum: [skill, connector_capability]
        target_ref:
          type: string
        schedule:
          type: string
          description: Supports cron descriptors and `@every <duration>`.
        enabled:
          type: boolean
        owner:
          type: string
        runtime_mode:
          type: string
        timeout_seconds:
          type: integer
        retry_max_attempts:
          type: integer
        retry_initial_backoff:
          type: string
        labels:
          type: object
          additionalProperties:
            type: string
        skill:
          $ref: "#/components/schemas/AutomationSkillTarget"
        connector_capability:
          $ref: "#/components/schemas/AutomationConnectorCapability"
        state:
          $ref: "#/components/schemas/AutomationJobState"
        last_run:
          $ref: "#/components/schemas/AutomationRun"
    AutomationListResponse:
      allOf:
        - $ref: "#/components/schemas/ListPage"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/AutomationJob"
    AutomationUpsertRequest:
      type: object
      required: [job]
      properties:
        job:
          $ref: "#/components/schemas/AutomationJob"
...